feat: Add streaming functionality to the language chain
This commit introduces a streaming functionality in the language chain. A new 'stream' function is added to the Chain trait that is expected to be implemented by all chains. A corresponding 'stream' implementation is provided in the LLMChain struct as well. The LLM trait has also been updated to include a 'stream' function. The 'stream' function is also implemented in the OpenAI struct, where the chat API is leveraged to create a stream of response generation.

In addition, a new example of usage of the streaming API has been included under the examples directory.