Feature/loader stream (#77)
* refactor: update CsvLoader to use BufReader for file reading; remove tokio dependencies; using buff read instead

* refactor: loader stream

* feat: Update document loaders to use async-stream

This update modifies the load functions within the document loaders (CsvLoader, LoPdfLoader, TextLoader) to make use of async-stream. The use of async-stream allows document loading to behave asynchronously which can potentially improve performance and resource usage especially when handling large sets of documents.

In addition, the tests are modified to map the results to handle the new asynchronous behaviour of the loaders. Two new dependencies, futures-util and async-stream, are also added to support this feature.

* refactor: extracting load and split functionality

extracting load and split functionality to and standalon function for
redability

* chore: changing test text to english