Feature/claude (#101)
* feat: Add cloude LLM

* chore(logs): logs

* chore(refactor): visibility of modules

* chore: update Cloude client and models to include default implementation and new functions

* chore: adding models, and chage model on api call

* feat: add error handling and payload customization for anthropic

This commit includes enhancements to the cloud API client. Error handling for different server responses has been added using the  enum added to . The  file has been extensively refactored to support this enhanced error handling and also offers better payload customization options. Default values for , , and  are now being set from the options. Tests have also been updated to reflect changes.

Additionally, a first class error handling module  has been created for handling different types of errors. This module contains the  struct for representing various types of errors that could occur when communicating with the Anthropic API. The  has been updated to include this new error module.

The models have been updated in  to support additional optional parameters in the payload like , , , and .

* chore: ignoring examples

* fix: fixing typo Cloude to Claude

* feat: improve content-type header and simplify check for streaming function in client.rs

This commit updates the content-type header to specify utf-8 charset. It also simplifies the way we check if a streaming function is present in the client's options.